+++ src/ps/plone/mlstiles/tests/test_setup.py
@@ -7,11 +7,8 @@ try:
 except ImportError:
     import unittest
 
-# zope imports
-from plone.registry.interfaces import IRegistry
-from zope.component import getUtility
-
 # local imports
+from ps.plone.mlstiles import PLONE_4
 from ps.plone.mlstiles.testing import PS_PLONE_MLSTILES_INTEGRATION_TESTING
 
 
@@ -47,31 +44,9 @@ class TestSetup(unittest.TestCase):
 
     def test_cssregistry(self):
         """Validate the CSS file registration."""
+        if not PLONE_4:
+            return
+
         resource_ids = self.portal.portal_css.getResourceIds()
         for id in CSS:
             self.assertIn(id, resource_ids, '{0} not installed'.format(id))
-
-    def test_tiles_registered(self):
-        """Validate that the tiles are registered."""
-        registry = getUtility(IRegistry)
-        key = 'plone.app.tiles'
-        self.assertIn(
-            'ps.plone.mlstiles.development_collection',
-            registry.records.get(key).value,
-        )
-        self.assertIn(
-            'ps.plone.mlstiles.listing_collection',
-            registry.records.get(key).value,
-        )
-        self.assertIn(
-            'ps.plone.mlstiles.featured_listings',
-            registry.records.get(key).value,
-        )
-        self.assertIn(
-            'ps.plone.mlstiles.recent_listings',
-            registry.records.get(key).value,
-        )
-        self.assertIn(
-            'ps.plone.mlstiles.listing_search',
-            registry.records.get(key).value,
-        )