+++ src/ps/plone/mlstiles/tiles/development_collection.py
@@ -1,51 +1,27 @@
 # -*- coding: utf-8 -*-
 """A tile that shows a list of MLS developments."""
 
+# python imports
+import copy
+
 # zope import
-from plone.supermodel.model import Schema
-from plone.tiles import Tile
+from plone import api as plone_api
+from plone.memoize import view
 from ps.plone.mls import (
-    # api,
+    api,
     config,
 )
+from ps.plone.mls.browser.developments.collection import (
+    EXCLUDED_SEARCH_FIELDS,
+    FIELDS,
+)
 from ps.plone.mls.interfaces import IDevelopmentCollection
-from zope import schema
 from zope.annotation.interfaces import IAnnotations
-from zope.schema.fieldproperty import FieldProperty
-
-# local imports
-from ps.plone.mlstiles import _
-from ps.plone.mlstiles.tiles.base import CatalogSource
+from zope.traversing.browser.absoluteurl import absoluteURL
 
 
-class IDevelopmentCollectionTile(Schema):
-    """Configuration schema for a development collection tile."""
-
-    count = schema.List(
-        required=False,
-        title=_(u'Number of items to display'),
-        value_type=schema.TextLine(),
-    )
-
-    offset = schema.Int(
-        default=0,
-        required=False,
-        title=_(u'Start at item'),
-    )
-
-    content_uid = schema.Choice(
-        title=_(u'Select an existing content'),
-        required=True,
-        source=CatalogSource(),
-    )
-
-
-class DevelopmentCollectionTile(Tile):
-    """A tile that shows a list of MLS developments."""
-
-    count = FieldProperty(IDevelopmentCollectionTile['count'])
-    offset = FieldProperty(IDevelopmentCollectionTile['offset'])
-    content_uid = FieldProperty(IDevelopmentCollectionTile['content_uid'])
+class DevelopmentCollectionTileMixin(object):
+    """A tile mixin that shows a list of MLS developments."""
 
     def get_config(self, obj):
         """Get collection configuration data from annotations."""
@@ -55,3 +31,74 @@ class DevelopmentCollectionTile(Tile):
     def has_development_collection(self, obj):
         """Check if the obj is activated for recent MLS developments."""
         return IDevelopmentCollection.providedBy(obj)
+
+    @property
+    def get_context(self):
+        """Return the development collection context."""
+        raise NotImplementedError
+
+    def get_fields(self):
+        return FIELDS
+
+    @property
+    def size(self):
+        raise NotImplementedError
+
+    @property
+    def start_at(self):
+        raise NotImplementedError
+
+    @property
+    @view.memoize
+    def items(self):
+        """Return the collection items."""
+        items = []
+        context = self.get_context
+        if not context or not self.has_development_collection(context):
+            return items
+
+        context_config = copy.copy(self.get_config(context))
+        language = plone_api.portal.get_current_language(context=context)
+        mlsapi = api.get_api(context=context, lang=language)
+        params = {
+            'fields': u','.join(self.get_fields()),
+            'limit': self.size,
+            'offset': self.start_at,
+        }
+        context_config.update(params)
+        params = api.prepare_search_params(
+            context_config,
+            context=context,
+            omit=EXCLUDED_SEARCH_FIELDS,
+        )
+        try:
+            result = api.Development.search(mlsapi, params=params)
+        except Exception:
+            return items
+        else:
+            items = result.get_items()
+        return items
+
+    def get_item_url(self, item):
+        """Get the (possibly modified) URL for the development item."""
+        config = {}
+        context = self.get_context
+        if context and self.has_development_collection(context):
+            config = copy.copy(self.get_config(context))
+
+        url = u'{0}{1}'.format(self.view_url(context), item.id.value)
+        if config.get('modify_url', True):
+            url = u'{0}___{1}-{2}'.format(
+                url,
+                item.title.value,
+                item.location.value,
+            )
+        return url
+
+    @view.memoize
+    def view_url(self, obj):
+        """Generate view url."""
+        return '/'.join([
+            absoluteURL(obj, self.request),
+            '',
+        ])